# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List prescriptions (admin/pharmacist see all; customers see their own)
      responses:
        "200":
          description: List of prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"
    post:
      operationId: createPrescription
      tags: [prescriptions]
      summary: Upload a prescription (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/{id}:
    get:
      operationId: getPrescription
      tags: [prescriptions]
      summary: Get prescription by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prescription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /prescriptions/{id}/verify:
    patch:
      operationId: verifyPrescription
      tags: [prescriptions]
      summary: Verify a prescription (pharmacist/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionReviewInput"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /prescriptions/{id}/reject:
    patch:
      operationId: rejectPrescription
      tags: [prescriptions]
      summary: Reject a prescription (pharmacist/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionReviewInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Orders ────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin/pharmacist see all; customers see their own)
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "400":
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order with line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (pharmacist/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Payments ──────────────────────────────────────────────────────────────
  /payments:
    post:
      operationId: createPayment
      tags: [payments]
      summary: Process payment for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Purchase Orders ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth ──────────────────────────────────────────────────────────────
...
    Prescription:
      type: object
      required: [id, customerId, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, verified, rejected]
        verifiedBy:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PrescriptionInput:
      type: object
      properties:
        imageUrl:
          type: string
        notes:
          type: string
    PrescriptionReviewInput:
      type: object
      properties:
        notes:
          type: string

    # ── Order ─────────────────────────────────────────────────────────────
    OrderItem:
      type: object
      required: [id, orderId, medicineId, quantity, price]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: string
    OrderItemInput:
      type: object
      required: [medicineId, quantity]
      properties:
        medicineId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    Order:
      type: object
      required: [id, customerId, status, total, paymentStatus, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        prescriptionId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, processing, dispensed, delivered, cancelled]
        subtotal:
          type: string
        total:
          type: string
        paymentStatus:
          type: string
          enum: [unpaid, paid, refunded]
        createdAt:
          type: string
          format: date-time
    OrderDetail:
      type: object
      required: [id, customerId, status, total, paymentStatus, items, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        prescriptionId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, processing, dispensed, delivered, cancelled]
        subtotal:
          type: string
        total:
          type: string
        paymentStatus:
          type: string
          enum: [unpaid, paid, refunded]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
    OrderInput:
      type: object
      required: [items]
      properties:
        prescriptionId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
          minItems: 1
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, processing, dispensed, delivered, cancelled]

    # ── Payment ───────────────────────────────────────────────────────────
    Payment:
      type: object
      required: [id, orderId, amount, method, status, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        amount:
          type: string
        method:
          type: string
          enum: [card, cash, insurance]
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        transactionId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PaymentInput:
      type: object
      required: [orderId, amount, method]
      properties:
        orderId:
          type: integer
        amount:
          type: string
        method:
          type: string
          enum: [card, cash, insurance]
        transactionId:
          type: string

    # ── Purchase Order ────────────────────────────────────────────────────
...