# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PrescriptionStatus } from './prescriptionStatus';

export interface Prescription {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  status: PrescriptionStatus;
  /** @nullable */
  verifiedBy?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}